==> notebooks/01 - Data Wrangling.ipynb <==
## Data Wrangling for Review Sentimental Analysis and Recommnadation System 

Importing the library

import pandas as pd
import numpy as np
import re

First import the review dataset and making it ready for the next step.

review = pd.read_csv('/Users/Amin/Documents/GitHub/Review-Sentiment-Analysis-with-Recommendation-System/data/reviews.csv')

review.head()

It looks a clean dataset, with name of the reviewer, name of the movie, rating that reviewer gets to film, summary and detail of the review and date of the review. 
First, there are some columns that are not related to the project goal and some need to do some changes.

review.describe()

review.isnull().sum()

review.dropna(inplace=True)

Some of the rows don't have the rating information. Becasue there is no way to fill these and there is enough data, those rows drop from the dataset.

review.reviewer.nunique()

After droping these NaN values, there are 11256 unique reviewer.

#checking the type of the data
review.dtypes

Next, the data type needs to be check and change to do some of the adjustment for the review, date, and rating columns.

#Droping the columns that are not related
review.drop(columns=['review_id','spoiler_tag','helpful','review_summary'],inplace=True)

review.head()

# Changing the type of the movie name and review details to a string.
review.review_detail = review.review_detail.astype('string')
review.movie = review.movie.astype('string')

review.dtypes

#Remove the comma and dot from the review
review.review_detail = review.review_detail.str.replace(',','').replace('.','')

#Makes all the review lowercase
review.review_detail = review.review_detail.str.lower()

#Change the type of the review date to datetime
review.review_date = pd.to_datetime(review.review_date)

review.dtypes

review.head()

The rating needs to change to a lable for sentiment analysis. Because I am going to make a recommandation system after, two lable chosed. " Love it" or "Not Love It". If the user love the movie we going to make a recommandation.

def conditions(review):
    if (review.rating > 7.0):
        return 'LOVE IT'

    else:
        return 'NOT LOVE IT'

review['lable'] = review.apply(conditions, axis=1)

#review.drop(columns=['rating'],inplace=True)
review.head()

review.lable.value_counts()

It seems there is a good distribution of the lable.

Next the movies dataset will load so we can add the imdb Id to the dataset so we can use the IMDB dataset to add more information of the movie to the data.

movie = pd.read_csv('/Users/Amin/Documents/GitHub/Review-Sentiment-Analysis-with-Recommendation-System/data/movies.csv')

movie.head()

movie['imdbId'] = movie['imdbId'].astype('string')
movie['imdb_id'] = 'tt' + movie['imdbId'].str.rjust(7,'0')

movie['Title'] = movie['Title'].str.replace(r"\(.*\)","")

movie.head()

#Merge the review and movie
df = review.merge(movie,left_on='movie',right_on='Title')

df.info()

The dataset Imdb Id format is diffrent from the current ID which has a 7 digit format with "tt" in the beginning.
Because we are going to use the ID to merge this data with IMDB data, this columns need to be adjust to the current standard. So we need to add "tt" and make it 7 digit but adding zero to the left.

#make the imdbId to string,Add the "tt" and adjust the number of digit
df['imdbId'] = df['imdbId'].astype('string')
df['imdb_id'] = 'tt' + df['imdbId'].str.rjust(7, '0')

df.head()

#The Movie title has the date at the end which needs to be remove.
df['movie'] = df['movie'].str.replace(r"\(.*\)","")

df.head()

df.drop(columns=['imdbId','Imdb Link','Title','IMDB Score','Genre','local_image_path'],inplace=True)

df.head()

df.to_csv('/Users/Amin/Documents/GitHub/Review-Sentiment-Analysis-with-Recommendation-System/data/review_db.csv',index=False)

#Load the IMDB dataset to add more information to the data
imdb_db = pd.read_csv('/Users/Amin/Documents/GitHub/Review-Sentiment-Analysis-with-Recommendation-System/data/IMDb movies.csv')

imdb_db.columns

#Merge the review data with imdb data on imdb ID
#imdb_db = df.merge(imdb_db,left_on='imdb_id',right_on='imdb_title_id')
imdb_db = imdb_db[imdb_db.imdb_title_id.isin(df.imdb_id)]
imdb_db.shape

imdb.head()

imdb['Title'] = imdb['Title'].str.replace(r"\(.*\)","")

imdb = pd.merge(imdb_db,movie.drop_duplicates(),right_on='imdb_id',left_on='imdb_title_id',how='inner')

imdb.shape

imdb.head()

For the sake of the hybrid recommandation, I am going to keep the name of the director and genre of the movie.

#Drop the unwanted columns
imdb.drop(columns=['imdbId','Imdb Link','IMDB Score','Genre','local_image_path',
                      'imdb_title_id','title','original_title','date_published','duration','country',
                      'language','writer','production_company','avg_vote','votes',
                      'budget','usa_gross_income','worlwide_gross_income','metascore','reviews_from_users',
                      'reviews_from_critics'],inplace=True)

imdb.drop(columns=['actors'],inplace=True)

imdb.to_csv('/Users/Amin/Documents/GitHub/Review-Sentiment-Analysis-with-Recommendation-System/data/imdb_subset.csv',index=False)

The genre columns include three value seperated by a comma. It needs to seperated to three columns for further use.

final_db[['genre_1','genre_2','genre_3']] = final_db['genre'].str.split(', ', expand=True)

final_db.drop(columns=['genre'],inplace=True)

Those movie that has only one genre, the other two columns will fill with the same genre.

final_db['genre_2'].fillna(final_db['genre_1'],inplace=True)

final_db['genre_3'].fillna(final_db['genre_2'],inplace=True)

final_db.head()

#Checking the final dataset type
final_db.dtypes

#number of unique reviewer
final_db.reviewer.nunique()

#Number of the unique movie
final_db.movie.nunique()

#The lable distribution
final_db.lable.value_counts()

#saving the clean data 
final_db.to_csv('/Users/Amin/Documents/GitHub/Review-Sentiment-Analysis-with-Recommendation-System/data/clean_db.csv',index=False)
