==> notebooks/04 - Making a Hybrid Movie Recommender.ipynb <==
# 04 - Making a Hybrid Movie Recommender:
#### Amin Khoeini

***

The purpose of this project is to get review from a user and in case that review is a positive one, the system recommand a movie to that user. we need to make recomandation engine that get a user id and a movie title and based on that recommand a list of the movie to that user. 

For that goals, we are going to make a hybrid filter. Makign a hybrid recommender consists of two step:


- The first step is to make a __content based filter__ to pick a list of movie close to the target movie.
    * 1 - We need to make a metadata of the movie avalible in the database.
    * 2 - Use year, director, actors, genre and description of the movie to make a metadata. For that we are going to remove any unwanted character from those, remove the stop words from the description and stemm the  description.
    * 3 - We add all of these columns to make a soup metadata column. We also add director columns three time to give it more weight.
    * 4 - Then we use tfidf vectorizer on that soup.
    * 5 - The final step is to use sklearn cosine similarity on the tfidf vector. This way we having a dataset of movie similarities based on thier content,genre, director and actors.
    * 6 - This way we can __pick 30 film that are close to the target film__ with our costum content_recommender function.
    
    
    
- Secound step is to make a __colaborative filter__, using a Surprise library. This way we can perdict scores for the 30 movie that produce by the previous filter for that target user. Finally,  we can sort that list based on the  stimated score and pick a top 10 as a final recomendation.

## Content Based Filter:

import pandas as pd
import numpy as np

from nltk.stem import SnowballStemmer
import nltk

from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import cosine_similarity

movies = pd.read_csv('/Users/Amin/Documents/GitHub/Review-Sentiment-Analysis-with-Recommendation-System/data/imdb_subset.csv')

movies.head()

movies.dtypes

First we need to change all the features to a string, and process them to be all lower case, remove the stop words from the description, commas from the genre and actors columns and add all them together.

# Remove the empty space at the end of the movie title
movies['Title'] = movies['Title'].str.strip()

# Change the data type to string
movies[['year', 'genre','director','actors','description']] = movies[['year', 'genre','director','actors','description']].astype('string')

movies.dtypes

# Some of the movie content are NaN so we just fill those with empty
movies = movies.fillna('')

# Remove the comma from the genre columns
movies['genre'] = movies['genre'].str.replace(","," ")
movies['genre'] = movies['genre'].str.lower()

# Remove the comma from the director columns for the movie that has more than one director
movies['director'] = movies['director'].str.replace(","," ")
movies['director'] = movies['director'].str.lower()

# Do the same for the actors columns
movies['actors'] = movies['actors'].str.replace(","," ")
movies['actors'] = movies['actors'].str.lower()

# lower case the description and remove the comma from it
movies['description'] = movies['description'].str.lower()
movies['description'] = movies['description'].str.replace(","," ")

# Remove the stop word from description and also do the stemmer on it to make it better for the vectorizer

stop = nltk.corpus.stopwords.words('english')

movies['description'] = movies['description'].apply(lambda x: ' '.join([word for word in x.split() if word not in (stop)]))

stemmer = SnowballStemmer('english')
w_tokenizer = nltk.tokenize.WhitespaceTokenizer()
def stemmer_text(text):
    return ' '.join([stemmer.stem(w) for w in w_tokenizer.tokenize(text)])

movies['description'] = movies['description'].apply(stemmer_text)

# Making a soup from all the columns, gives director name more weight by adding it twice to the soup
movies['soup'] = movies['year'] +' '+ movies ['director'] + ' '+ movies ['director'] +' '+ movies['genre'] + ' '+movies ['description'] +' '+ movies ['actors']

movies.soup[0]

# Use CountVectorizer on the soup columns

count = CountVectorizer(analyzer='word',ngram_range=(1, 2),min_df=0, stop_words='english')
count_matrix = count.fit_transform(movies['soup'])

# Make a cosin similarity matrix on the countvectorizer so we can calculate the similarity of the movie to eachother based on the content
cosine_sim = cosine_similarity(count_matrix, count_matrix)

# Create a dataframe of that similarity matrix
movie_sim = pd.DataFrame(cosine_sim, index=movies.Title, columns=movies.Title)

def content_recommender(movie_title):
    ''' This function uses the similarity matrix that we create based on the count vector of the movie content
     and create a list of 30 movie that are similar to the target movie'''
    
    # Selecting the target movie similarity matrix
    cosine_similarity_series = movie_sim.loc[movie_title]
    
    # Sort these values highest to lowest and pick the first 30 movie.
    ordered_similarities = cosine_similarity_series.sort_values(ascending=False)[1:31]
    # 
    return (ordered_similarities.index.tolist())

# Get the recommandation for the Gangs of New York, we can see that the director name has lot of influence,
# The top recommandation of from same director.
content_recommender('Superman IV: The Quest for Peace (1987)')

## Collaborative Filter by Surprise:

The content based filter is only capable of suggesting movies based on its content and are not capable of capturing tastes of the viewer and how they might vote for the certin film.With this filter all the user that like a certain movie will get a same recommandation.

To enhance our recommandation, we will use Collaborative Filtering to make recommendations to users. Collaborative Filtering is based on the idea that users similar to eachother can be used to predict how much certain user will like a particular movie.

Surprise library provide extremely powerful algorithms to minimise RMSE (Root Mean Square Error) and give great recommendations.

at first step, we try most of these algorithms on our dataset to see which one has the least RMSE.

from surprise import SVD,SVDpp,SlopeOne, NMF, NormalPredictor, KNNBaseline, KNNBasic,KNNWithMeans, KNNWithZScore, BaselineOnly
from surprise.model_selection import cross_validate,GridSearchCV,train_test_split
import surprise
from surprise import Dataset
from surprise import Reader
from surprise import accuracy
import csv

# import the rating data 
ratings = pd.read_csv('/Users/Amin/Documents/GitHub/Review-Sentiment-Analysis-with-Recommendation-System/data/process_db.csv')

# Drop the review and only keep the user_id,movie_id and vote number
ratings.drop(columns=['review_detail','review_clean','lable','review_date','reviewer'],inplace=True)

# Make a instant of reader and data which is neccesary to make a model in Surprise

reader = Reader(rating_scale=(1, 10))
data = Dataset.load_from_df(ratings[['User_ID', 'imdb_id', 'rating']], reader)

Now that everything is ready to train a Surprise model, it is time to use a crossvalidation and get the performance of each of the surprise algorithms on our dataset.

benchmark = []
# Iterate over all algorithms
algorithms = [SVD(), SVDpp(), SlopeOne(), NMF(), NormalPredictor(), KNNBaseline(), KNNBasic(),
 KNNWithMeans(), KNNWithZScore(), BaselineOnly()]

for algorithm in algorithms:
    results = cross_validate(algorithm, data, measures=['RMSE'], cv=3, verbose=False,n_jobs=-1)
 
 # Get results & append algorithm name
    tmp = pd.DataFrame.from_dict(results).mean(axis=0)
    tmp = tmp.append(pd.Series([str(algorithm).split(' ')[0].split('.')[-1]], index=['Algorithm']))
    benchmark.append(tmp)

performance = pd.DataFrame(benchmark)
performance

__KNNBaseline has the least RMSE__ amoung the algorithms, although it has one the longest performance time but it will pick as our best model.
Next, we will try to see if we can get a better performance by doing a hyperparameters tunning.

trainset, testset = train_test_split(data, test_size=0.2)

# HyperParameters Tunning For sgd method

list_of_ks = [10,20,40]

sgd_bsl_options = [
    {'method':'sgd', 'reg': 0.02, 'learning_rate': 0.005},
    {'method':'sgd', 'reg': 0.05, 'learning_rate': 0.005},
    {'method':'sgd', 'reg': 0.1, 'learning_rate': 0.005},
    {'method':'sgd', 'reg': 0.02, 'learning_rate': 0.01},
    {'method':'sgd', 'reg': 0.05, 'learning_rate': 0.01},
    {'method':'sgd', 'reg': 0.1, 'learning_rate': 0.01}]


with open('/Users/Amin/Documents/GitHub/Review-Sentiment-Analysis-with-Recommendation-System/kNN_baseline_sgd_scores.csv', 'w') as f:
    writer = csv.writer(f)
    writer.writerow(['reg', 'learning_rate', 'k', 'train_rmse', 'test_rmse'])

res=[]
for curr_bsl_option in sgd_bsl_options:

    for curr_k in list_of_ks:
        
        print(
            'Currently calculating k = ' + str(curr_k) + ' ...'
        )        
        algo = KNNBaseline(k = curr_k, bsl_options = curr_bsl_option)
        results = cross_validate(algo, data, measures=['RMSE'], cv=3, return_train_measures=True);
        
        with open('/Users/Amin/Documents/GitHub/Review-Sentiment-Analysis-with-Recommendation-System/kNN_baseline_sgd_scores.csv', 'a') as f:
            writer = csv.writer(f)
            writer.writerow(
                [curr_bsl_option['reg'], curr_bsl_option['learning_rate'],str(curr_k), 
                 str(np.mean(results['train_rmse'])), str(np.mean(results['test_rmse']))])


df = pd.read_csv('/Users/Amin/Documents/GitHub/Review-Sentiment-Analysis-with-Recommendation-System/kNN_baseline_sgd_scores.csv')
df.sort_values(by = 'test_rmse', inplace = True)
df

# HyperParameters Tunning For als method

als_bsl_options = [
    {'method':'als', 'reg_i': 20, 'reg_u': 30},
    {'method':'als', 'reg_i': 40, 'reg_u': 60},
    {'method':'als', 'reg_i': 20, 'reg_u': 30},
    {'method':'als', 'reg_i': 40, 'reg_u': 60}
]

with open('/Users/Amin/Documents/GitHub/Review-Sentiment-Analysis-with-Recommendation-System/kNN_baseline_als_scores.csv', 'w') as f:
    writer = csv.writer(f)
    writer.writerow(['reg_i', 'reg_u', 'k', 'train_rmse', 'test_rmse'])


for curr_bsl_option in als_bsl_options:

    for curr_k in list_of_ks:
        
        print(
            'Currently calculating k = ' + str(curr_k) + ' ...'
        )        
        algo = KNNBaseline(k = curr_k, bsl_options = curr_bsl_option)
        results = cross_validate(algo, data, measures=['RMSE'], cv=3, return_train_measures=True);
        
        with open('/Users/Amin/Documents/GitHub/Review-Sentiment-Analysis-with-Recommendation-System/kNN_baseline_als_scores.csv', 'a') as f:
            writer = csv.writer(f)
            writer.writerow(
                [curr_bsl_option['reg_i'], curr_bsl_option['reg_u'],str(curr_k), 
                 str(np.mean(results['train_rmse'])), str(np.mean(results['test_rmse']))])

df1 = pd.read_csv('/Users/Amin/Documents/GitHub/Review-Sentiment-Analysis-with-Recommendation-System/kNN_baseline_als_scores.csv')
df1.sort_values(by = 'test_rmse', inplace = True)
df1

The Hyperparameters tunning lower the RMSE by 0.2, while we train the model Using Stochastic Gradient Descent method with learning rate of 0.01, 40 neighbors and regularization parameter set to 0.1

# Making a Best Suprise Model for our data
model = KNNBaseline(k = 40, bsl_options = {'method':'sgd', 'reg': 0.1, 'learning_rate': 0.01})

# Train The Model on Full Dataset 
full_trainset = data.build_full_trainset()
model.fit(full_trainset)


Now by giving the user_id and movie_id to the model we can predict the how that user would vote that movie. 

We will use this model in the next step which is combining the two filter and create our final recommandation engine.

# Get the prediction for user_id 9090 and movie_id tt0208988
model.predict(uid=9090, iid='tt0208988')

est=[]
movies_list = movies_db
ids =  indices['imdb_id']
for b in ids:
    stm = model.predict(2,b).est
    est.append(stm)
movies_list['est'] = est
    
    # Add the estimated vote from collaborative filter and add it to our suggested movie dataset, and sort the data based on that estimated vote
movies_list = movies_list.sort_values('est', ascending=False)
    # Return the top 10 as our final movie recommandation

movies_list[0:20]

## Making a Final Recommandation System with Hybrid Filter:

In the final section, we will try to build a simple hybrid recommender that brings together what we have implemented in the content based and collaborative filter based engines. 

This final filter will work like this:

- Input: User ID and the Title of a Movie
- Output: Similar movies sorted on the basis of expected ratings by that particular user.



# Create a index list for the movie title and movie id and set the title as a index
indices = movies[['Title','imdb_id']].set_index('Title')

# make a subset of the movie metadata and set the index to Title
movies_db = movies[['Title','director','year','genre']].set_index('Title')

# create our final recommandation engine 
def recommender(user_id,movie_title):
    
    ''' This function used the two previous filter and create the final recommandation movie list.
    First it creates the list of 50 movie by using the content based filter.
    Secound, using the suprise model, predict the vote that target user would gives to those movie.
    Third sort the 50 movie based on that predicted vote
    Finally, it return the top 10 movies as a final recommandation
    
    '''
    
    est=[]
    # First use the content based filter to make a list of movie that are close to the target movie
    rec_movies = content_recommender(movie_title)
    # subset the movie metadata to have only movie that the content base filter suggested
    movies_list = movies_db.loc[rec_movies]
    # Get the movie_id of the suggested movie
    ids =  indices.loc[rec_movies]['imdb_id']
    # Using the surprise model and collaborative filtering to predict the vote that the target user gives to the suggested movie
    for b in ids:
        stm = model.predict(user_id,b).est
        est.append(stm)
    movies_list['est'] = est
    
    # Add the estimated vote from collaborative filter and add it to our suggested movie dataset, and sort the data based on that estimated vote
    movies_list = movies_list.sort_values('est', ascending=False)
    # Return the top 10 as our final movie recommandation
    return movies_list[0:10]

recommender(2,'Gangs of New York (2002)')

recommender(1,'Kill Bill: Vol. 2 (2004)')

recommender(2344,'Kill Bill: Vol. 2 (2004)')
